backend/DermaAI/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dermaai"
version = "1.1.0"
description = "DermaAI CLI - Skin Lesion Detection Tool"
authors = [{name="Ngane Emmanuel"}]
readme = "README.md"
requires-python = ">=3.10"
license = {text="MIT"}

dependencies = [
    "pillow~=11.3.0",
    "numpy~=2.3.2",
    "pandas~=2.3.1",
    "torch~=2.7.1",
    "torchvision~=0.22.1",
    "typer[all]~=0.16.1",
    "pypandoc~=1.15"
]


[project.scripts]
dermai = "dermaai_cli.cli:main"


